Resolved bug with pagination.
Pagination had bug when the number of documents in a page of listing of documents in a collection was not a factor of total number of documents in the collection. Resolved it.

+++ mongonaut/views.py
@@ -96,8 +96,10 @@ class DocumentListView(MongonautViewMixin, FormView):
             self.page = 1
 
         obj_count = queryset.count()
-        self.total_pages = obj_count / self.documents_per_page + (1 if obj_count % self.documents_per_page else 0)
-
+        self.total_pages = int(
+            obj_count / self.documents_per_page + (1 if obj_count % self.documents_per_page else 0)
+            )
+            
         if self.page < 1:
             self.page = 1
